+++ History.md
@@ -1,4 +1,8 @@
-# Changelog for chefdk-template-example
+# Changelog for chef-gen-flavor-example
+
+## 0.2.0
+
+* rename from chefdk-template-example to chef-gen-flavor-example
 
 ## 0.1.0